+++ demo/src/app/shared/index.ts
@@ -1,1 +1,1 @@
-export * from './share.module';
+export * from './shared.module';