+++ src/config/gallery-thumb.config.ts
@@ -1,6 +1,0 @@
-export interface GalleryThumbConfig {
-    width?: number;
-    height?: number;
-    position?: string;
-    space?: number;
-}