+++ src/models/index.ts
@@ -1,0 +1,15 @@
+export {
+  GalleryConfig,
+  GalleryDescConfig,
+  GalleryThumbConfig,
+  GalleryNavConfig,
+  GalleryLoaderConfig,
+  GalleryPlayConfig,
+  GalleryBulletConfig,
+  GalleryLightboxConfig
+} from './gallery.config';
+
+export {
+  GalleryState,
+  GalleryItem
+} from './gallery.models';