+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.0.0
+
+ - fix(keyboard listener in lightbox) closes [#24](https://github.com/MurhafSousli/ng-gallery/issues/24), [#33](https://github.com/MurhafSousli/ng-gallery/issues/33).
+ - refactor(Gallerize directive) Use MutationObserver instead of DOMSubtreeModified, closes [#26](https://github.com/MurhafSousli/ng-gallery/issues/26).
+ - fix(Universal support), closes [#9](https://github.com/MurhafSousli/ng-gallery/issues/9).
+ - fix Angular 5 warning, closes [#21](https://github.com/MurhafSousli/ng-gallery/issues/21).
+ - Improve gallery lightbox, closes [#20](https://github.com/MurhafSousli/ng-gallery/issues/20).
+ - Improve gallery lightbox slide animation, closes [#8](https://github.com/MurhafSousli/ng-gallery/issues/8).
+ - Use Angular CDK for the gallery lightbox.
+ - refactor(GalleryConfig)
+ - Remove image transition animation option because it was not implemented properly.
+
 ## 0.7.1
 
  - General refactor